Remove fluxconfig env checks from webhook validation, update cli preflight check for gitops
