+++ pkg/validations/createvalidations/gitops.go
@@ -5,21 +5,13 @@ import (
 	"fmt"
 
 	"github.com/aws/eks-anywhere/pkg/cluster"
-	"github.com/aws/eks-anywhere/pkg/config"
 	"github.com/aws/eks-anywhere/pkg/logger"
 	"github.com/aws/eks-anywhere/pkg/types"
 	"github.com/aws/eks-anywhere/pkg/validations"
 )
 
-func ValidateGitOps(ctx context.Context, k validations.KubectlClient, cluster *types.Cluster, spec *cluster.Spec, cliConfig *config.CliConfig) error {
-	if spec.FluxConfig != nil {
-		err := validateAuthenticationGitProvider(spec, cliConfig)
-		if err != nil {
-			return err
-		}
-	}
-
-	if spec.GitOpsConfig == nil || spec.Cluster.IsSelfManaged() {
+func ValidateGitOps(ctx context.Context, k validations.KubectlClient, cluster *types.Cluster, spec *cluster.Spec) error {
+	if spec.GitOpsConfig == nil {
 		logger.V(5).Info("skipping ValidateGitOps")
 		return nil
 	}
@@ -38,36 +30,6 @@ func ValidateGitOps(ctx context.Context, k validations.KubectlClient, cluster *t
 	}
 
 	return err
-}
-
-func validateAuthenticationGitProvider(clusterSpec *cluster.Spec, cliConfig *config.CliConfig) error {
-	fluxConfig := clusterSpec.FluxConfig
-	if fluxConfig.Spec.Git == nil {
-		return nil
-	}
-
-	if cliConfig.GitPrivateKeyFile == "" {
-		return fmt.Errorf("provide a path to a private key file via the EKSA_GIT_PRIVATE_KEY " +
-			"in order to use the generic git Flux provider")
-	}
-
-	if cliConfig.GitPrivateKeyFile != "" {
-		if !validations.FileExistsAndIsNotEmpty(cliConfig.GitPrivateKeyFile) {
-			return fmt.Errorf("private key file does not exist at %s or is empty", cliConfig.GitPrivateKeyFile)
-		}
-	}
-
-	if cliConfig.GitKnownHostsFile == "" {
-		return fmt.Errorf("provide a path to an SSH Known Hosts file which contains a valid entry associate with the given private key via the EKSA_GIT_SSH_KNOWN_HOSTS environment variable")
-	}
-
-	if cliConfig.GitKnownHostsFile != "" {
-		if !validations.FileExistsAndIsNotEmpty(cliConfig.GitKnownHostsFile) {
-			return fmt.Errorf("SSH known hosts file does not exist at %v or is empty", cliConfig.GitKnownHostsFile)
-		}
-	}
-
-	return nil
 }
 
 func validateWorkloadFields(ctx context.Context, k validations.KubectlClient, cluster *types.Cluster, spec *cluster.Spec) error {
