+++ pkg/validations/createvalidations/gitops_test.go
@@ -10,14 +10,8 @@ import (
 	"github.com/aws/eks-anywhere/internal/test"
 	"github.com/aws/eks-anywhere/pkg/api/v1alpha1"
 	"github.com/aws/eks-anywhere/pkg/cluster"
-	"github.com/aws/eks-anywhere/pkg/config"
 	"github.com/aws/eks-anywhere/pkg/validations"
 	"github.com/aws/eks-anywhere/pkg/validations/createvalidations"
-)
-
-const (
-	emptyVar   = ""
-	testEnvVar = "test"
 )
 
 var eksaGitOpsResourceType = fmt.Sprintf("gitopsconfigs.%s", v1alpha1.GroupVersion.Group)
@@ -137,7 +131,6 @@ func TestValidateGitopsForWorkloadClustersPath(t *testing.T) {
 				s.GitOpsConfig = defaultGitOps
 				s.Cluster.SetManagedBy("management-cluster")
 			})
-			cliConfig := &config.CliConfig{}
 			k, ctx, cluster, e := validations.NewKubectl(t)
 			cluster.Name = "management-cluster"
 
@@ -161,7 +154,7 @@ func TestValidateGitopsForWorkloadClustersPath(t *testing.T) {
 					cluster.KubeconfigFile, "--namespace", clusterSpec.Cluster.Namespace,
 				}).Return(*bytes.NewBufferString(mgmtGitOpsContent), nil)
 
-			err := createvalidations.ValidateGitOps(ctx, k, cluster, clusterSpec, cliConfig)
+			err := createvalidations.ValidateGitOps(ctx, k, cluster, clusterSpec)
 			if !reflect.DeepEqual(err, tc.wantErr) {
 				t.Errorf("%v got = %v, \nwant %v", tc.name, err, tc.wantErr)
 			}
@@ -206,7 +199,6 @@ func TestValidateGitopsForWorkloadClustersFailure(t *testing.T) {
 		s.GitOpsConfig = defaultGitOps
 		s.Cluster.SetManagedBy("management-cluster")
 	})
-	cliConfig := &config.CliConfig{}
 	k, ctx, cluster, e := validations.NewKubectl(t)
 	cluster.Name = testclustername
 	for _, tc := range tests {
@@ -219,7 +211,7 @@ func TestValidateGitopsForWorkloadClustersFailure(t *testing.T) {
 					"--field-selector=metadata.name=gitopstest",
 				}).Return(*bytes.NewBufferString(fileContent), nil)
 
-			err := createvalidations.ValidateGitOps(ctx, k, cluster, clusterSpec, cliConfig)
+			err := createvalidations.ValidateGitOps(ctx, k, cluster, clusterSpec)
 			if !reflect.DeepEqual(err, tc.wantErr) {
 				t.Errorf("%v got = %v, \nwant %v", tc.name, err, tc.wantErr)
 			}
@@ -244,7 +236,6 @@ func TestSkipValidateGitopsWithNoGitOpts(t *testing.T) {
 
 		s.GitOpsConfig = nil
 	})
-	cliConfig := &config.CliConfig{}
 
 	k, ctx, cluster, e := validations.NewKubectl(t)
 	cluster.Name = testclustername
@@ -257,7 +248,7 @@ func TestSkipValidateGitopsWithNoGitOpts(t *testing.T) {
 					"--field-selector=metadata.name=gitopstest",
 				}).Times(0)
 
-			err := createvalidations.ValidateGitOps(ctx, k, cluster, clusterSpec, cliConfig)
+			err := createvalidations.ValidateGitOps(ctx, k, cluster, clusterSpec)
 			if !reflect.DeepEqual(err, tc.wantErr) {
 				t.Errorf("%v got = %v, \nwant %v", tc.name, err, tc.wantErr)
 			}
@@ -267,29 +258,12 @@ func TestSkipValidateGitopsWithNoGitOpts(t *testing.T) {
 
 func TestValidateGitopsForSelfManagedCluster(t *testing.T) {
 	tests := []struct {
-		name               string
-		wantErr            error
-		getClusterResponse string
+		name    string
+		wantErr error
 	}{
 		{
-			name:               "Skip Validate GitOpsConfig name",
-			wantErr:            nil,
-			getClusterResponse: "testdata/empty_get_gitops_response.json",
-		},
-	}
-
-	defaultGitOps := &v1alpha1.GitOpsConfig{
-		Spec: v1alpha1.GitOpsConfigSpec{
-			Flux: v1alpha1.Flux{
-				Github: v1alpha1.Github{
-					Owner:               "owner",
-					Repository:          "repo",
-					FluxSystemNamespace: "flux-system",
-					Branch:              "main",
-					ClusterConfigPath:   "clusters/" + testclustername,
-					Personal:            false,
-				},
-			},
+			name:    "Skip ValidateGitOps",
+			wantErr: nil,
 		},
 	}
 
@@ -299,151 +273,15 @@ func TestValidateGitopsForSelfManagedCluster(t *testing.T) {
 			Kind: v1alpha1.GitOpsConfigKind,
 			Name: "gitopstest",
 		}
-		s.GitOpsConfig = defaultGitOps
+		s.GitOpsConfig = nil
 
 		s.Cluster.SetSelfManaged()
 	})
-	cliConfig := &config.CliConfig{}
-	k, ctx, cluster, e := validations.NewKubectl(t)
+	k, ctx, cluster, _ := validations.NewKubectl(t)
 	cluster.Name = testclustername
 	for _, tc := range tests {
 		t.Run(tc.name, func(tt *testing.T) {
-			e.EXPECT().Execute(
-				ctx, []string{
-					"get", eksaGitOpsResourceType, "-o", "json", "--kubeconfig",
-					cluster.KubeconfigFile, "--namespace", clusterSpec.Cluster.Namespace,
-					"--field-selector=metadata.name=gitopstest",
-				}).Times(0)
-
-			err := createvalidations.ValidateGitOps(ctx, k, cluster, clusterSpec, cliConfig)
-			if !reflect.DeepEqual(err, tc.wantErr) {
-				t.Errorf("%v got = %v, \nwant %v", tc.name, err, tc.wantErr)
-			}
-		})
-	}
-}
-
-func TestValidateGitOpsGitProviderNoAuthForWorkloadCluster(t *testing.T) {
-	tests := []struct {
-		name      string
-		wantErr   error
-		git       *v1alpha1.GitProviderConfig
-		cliConfig *config.CliConfig
-	}{
-		{
-			name:    "Empty password and private key",
-			wantErr: fmt.Errorf("provide a path to a private key file via the EKSA_GIT_PRIVATE_KEY in order to use the generic git Flux provider"),
-			git: &v1alpha1.GitProviderConfig{
-				RepositoryUrl: "testRepo",
-			},
-			cliConfig: &config.CliConfig{
-				GitPrivateKeyFile:   emptyVar,
-				GitSshKeyPassphrase: emptyVar,
-				GitKnownHostsFile:   "testdata/git_nonempty_ssh_known_hosts",
-			},
-		},
-		{
-			name:      "Empty git config",
-			wantErr:   nil,
-			git:       nil,
-			cliConfig: nil,
-		},
-		{
-			name:    "Empty password",
-			wantErr: fmt.Errorf("private key file does not exist at %s or is empty", testEnvVar),
-			git: &v1alpha1.GitProviderConfig{
-				RepositoryUrl: "testRepo",
-			},
-			cliConfig: &config.CliConfig{
-				GitPrivateKeyFile:   testEnvVar,
-				GitSshKeyPassphrase: emptyVar,
-				GitKnownHostsFile:   "testdata/git_nonempty_ssh_known_hosts",
-			},
-		},
-		{
-			name:    "Empty private key file",
-			wantErr: fmt.Errorf("private key file does not exist at %s or is empty", "testdata/git_empty_file"),
-			git: &v1alpha1.GitProviderConfig{
-				RepositoryUrl: "testRepo",
-			},
-			cliConfig: &config.CliConfig{
-				GitPrivateKeyFile:   "testdata/git_empty_file",
-				GitSshKeyPassphrase: emptyVar,
-				GitKnownHostsFile:   "testdata/git_nonempty_ssh_known_hosts",
-			},
-		},
-		{
-			name:    "Empty private key",
-			wantErr: fmt.Errorf("provide a path to a private key file via the EKSA_GIT_PRIVATE_KEY in order to use the generic git Flux provider"),
-			git: &v1alpha1.GitProviderConfig{
-				RepositoryUrl: "testRepo",
-			},
-			cliConfig: &config.CliConfig{
-				GitPrivateKeyFile:   emptyVar,
-				GitSshKeyPassphrase: testEnvVar,
-				GitKnownHostsFile:   "testdata/git_nonempty_ssh_known_hosts",
-			},
-		},
-		{
-			name:    "Password and private key populated",
-			wantErr: nil,
-			git: &v1alpha1.GitProviderConfig{
-				RepositoryUrl: "testRepo",
-			},
-			cliConfig: &config.CliConfig{
-				GitPrivateKeyFile:   "testdata/git_nonempty_private_key",
-				GitSshKeyPassphrase: testEnvVar,
-				GitKnownHostsFile:   "testdata/git_nonempty_ssh_known_hosts",
-			},
-		},
-		{
-			name:    "Empty known hosts",
-			wantErr: fmt.Errorf("SSH known hosts file does not exist at testdata/git_empty_file or is empty"),
-			git: &v1alpha1.GitProviderConfig{
-				RepositoryUrl: "testRepo",
-			},
-			cliConfig: &config.CliConfig{
-				GitPrivateKeyFile:   "testdata/git_nonempty_private_key",
-				GitSshKeyPassphrase: testEnvVar,
-				GitKnownHostsFile:   "testdata/git_empty_file",
-			},
-		},
-		{
-			name:    "No known hosts",
-			wantErr: fmt.Errorf("SSH known hosts file does not exist at testdata/git_empty_file or is empty"),
-			git: &v1alpha1.GitProviderConfig{
-				RepositoryUrl: "testRepo",
-			},
-			cliConfig: &config.CliConfig{
-				GitPrivateKeyFile:   "testdata/git_nonempty_private_key",
-				GitSshKeyPassphrase: testEnvVar,
-				GitKnownHostsFile:   "testdata/git_empty_file",
-			},
-		},
-	}
-
-	for _, tc := range tests {
-		t.Run(tc.name, func(tt *testing.T) {
-			defaultFlux := &v1alpha1.FluxConfig{
-				Spec: v1alpha1.FluxConfigSpec{
-					Git: tc.git,
-				},
-			}
-			clusterSpec := test.NewClusterSpec(func(s *cluster.Spec) {
-				s.Cluster.Name = testclustername
-				s.Cluster.Spec.GitOpsRef = &v1alpha1.Ref{
-					Kind: v1alpha1.GitOpsConfigKind,
-					Name: "gitopstest",
-				}
-				s.FluxConfig = defaultFlux
-				s.Cluster.SetManagedBy("management-cluster")
-			})
-
-			cliConfig := tc.cliConfig
-			k, ctx, cluster, _ := validations.NewKubectl(t)
-			cluster.Name = "management-cluster"
-
-			err := createvalidations.ValidateGitOps(ctx, k, cluster, clusterSpec, cliConfig)
+			err := createvalidations.ValidateGitOps(ctx, k, cluster, clusterSpec)
 			if !reflect.DeepEqual(err, tc.wantErr) {
 				t.Errorf("%v got = %v, \nwant %v", tc.name, err, tc.wantErr)
 			}
