+++ pkg/validations/createvalidations/preflightvalidations.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 
 	anywherev1 "github.com/aws/eks-anywhere/pkg/api/v1alpha1"
+	"github.com/aws/eks-anywhere/pkg/config"
 	"github.com/aws/eks-anywhere/pkg/features"
 	"github.com/aws/eks-anywhere/pkg/types"
 	"github.com/aws/eks-anywhere/pkg/validations"
@@ -30,6 +31,11 @@ func (u *CreateValidations) PreflightValidations(ctx context.Context) (err error
 			Err:         validations.ValidateK8s123Support(u.Opts.Spec),
 			Silent:      true,
 		},
+		{
+			Name:        "validate authentication for git provider",
+			Remediation: fmt.Sprintf("ensure %s, %s env variable are set and valid", config.EksaGitPrivateKeyTokenEnv, config.EksaGitKnownHostsFileEnv),
+			Err:         validations.ValidateAuthenticationForGitProvider(u.Opts.Spec, u.Opts.CliConfig),
+		},
 	}
 
 	if u.Opts.Spec.Cluster.IsManaged() {
@@ -43,7 +49,7 @@ func (u *CreateValidations) PreflightValidations(ctx context.Context) (err error
 			validations.ValidationResult{
 				Name:        "validate gitops",
 				Remediation: "",
-				Err:         ValidateGitOps(ctx, k, u.Opts.ManagementCluster, u.Opts.Spec, u.Opts.CliConfig),
+				Err:         ValidateGitOps(ctx, k, u.Opts.ManagementCluster, u.Opts.Spec),
 			},
 			validations.ValidationResult{
 				Name:        "validate identity providers' name",
