+++ pkg/validations/testdata/git_nonempty_private_key
